requirements.txt
SpeechRecognition
arabic_reshaper
python-bidi
gTTS
pygame
cohere
